refactor: extract CodeAnalysis signature
Separates the CodeAnalysis DSPy signature into its own dedicated module.
