+++ src/robofactor/dspy_modules.py
@@ -3,21 +3,11 @@ from typing import Literal
 
 import dspy
 
+from robofactor.signatures.code_analysis import CodeAnalysis
+
 FAILURE_SCORE = 0.0
 TRAINING_DATA_FILE = "training.json"
 logger = logging.getLogger(__name__)
-
-
-class CodeAnalysis(dspy.Signature):
-    """
-    Analyze Python code for its purpose, complexity, and dependencies.
-    Identify actionable refactoring opportunities and summarize the findings.
-    """
-    code_snippet: dspy.Code[Literal["python"]] = dspy.InputField(desc="Python code to analyze")
-    analysis: str = dspy.OutputField(description="Summary of functionality, complexity, and dependencies")
-    refactoring_opportunities: list[str] = dspy.OutputField(
-        description="Actionable bullet points for refactoring"
-    )
 
 
 class RefactoringPlan(dspy.Signature):
